Auto-KB: Adding more automation functions
This PR adds

- update
- review
- apply-patch

commands and operations

update allows individual fields to be updated; for example, if
only some of your records have `description`, this will populate others
(using RAG)

the review command will ask the LLM to be a critic, and evaluate
individual assertions using its own latent KB, guided by the curator.

A new output format is supported, "patch" which makes it easier
to use these commands as part of a curation cycle

It also enhances the citeseek function to allow for weaving in multiple pieces
of evidence (dynamically retrieved from pubmed) to complex objects.

Also adds two wrappers

- fairsharing
- jgi data portal